RFR: webview to its own file - add padding for logout button - RMV: focus metric
